+++ src/main/java/subway/path/util/LoginMemberAgeFare.java
@@ -1,4 +1,4 @@
-package subway.path.domain;
+package subway.path.util;
 
 import subway.member.domain.LoginMember;
